+++ common/operations/Slice.cpp
@@ -44,7 +44,7 @@ namespace {
 
 template <typename T>
 void addVectors(const std::vector<T>& a, const std::vector<T>& b, std::vector<T>* res) {
-    for (int i = 0; i < res->size(); ++i) {
+    for (size_t i = 0; i < res->size(); ++i) {
         res->at(i) = a[i] + b[i];
     }
 }
@@ -60,7 +60,7 @@ bool evalGeneric(const T* inputData, const Shape& inputShape, const int32_t* beg
     std::vector<uint32_t> beginIndex(getSizeOfDimension(beginShape, 0));
     std::vector<uint32_t> inputIndex(getNumberOfDimensions(inputShape));
 
-    for (int i = 0; i < beginIndex.size(); ++i) {
+    for (size_t i = 0; i < beginIndex.size(); ++i) {
         beginIndex[i] = static_cast<uint32_t>(beginData[i]);
     }
 
@@ -109,15 +109,15 @@ Result<Version> validate(const IOperationValidationContext* context) {
 #ifdef NN_INCLUDE_CPU_IMPLEMENTATION
 bool prepare(IOperationExecutionContext* context) {
     const Shape& inputShape = context->getInputShape(kInputTensor);
-    const int32_t n_dims = getNumberOfDimensions(inputShape);
+    const uint32_t n_dims = getNumberOfDimensions(inputShape);
     NN_RET_CHECK(n_dims > 0);
 
     const Shape& beginShape = context->getInputShape(kBeginTensor);
-    NN_RET_CHECK_EQ(getNumberOfDimensions(beginShape), 1);
+    NN_RET_CHECK_EQ(getNumberOfDimensions(beginShape), 1u);
     NN_RET_CHECK_EQ(getSizeOfDimension(beginShape, 0), n_dims);
 
     const Shape& sizeShape = context->getInputShape(kSizeTensor);
-    NN_RET_CHECK_EQ(getNumberOfDimensions(sizeShape), 1);
+    NN_RET_CHECK_EQ(getNumberOfDimensions(sizeShape), 1u);
     NN_RET_CHECK_EQ(getSizeOfDimension(sizeShape, 0), n_dims);
 
     const int32_t* beginData = context->getInputBuffer<int32_t>(kBeginTensor);
@@ -125,15 +125,16 @@ bool prepare(IOperationExecutionContext* context) {
 
     Shape outputShape = context->getOutputShape(kOutputTensor);
     outputShape.dimensions.resize(n_dims);
-    for (int i = 0; i < n_dims; ++i) {
+    for (uint32_t i = 0; i < n_dims; ++i) {
         const int32_t sliceBegin = beginData[i];
         int32_t sliceSize = sizeData[i];
         if (sliceSize == -1) {
             sliceSize = getSizeOfDimension(inputShape, i) - sliceBegin;
         }
-        NN_RET_CHECK_LE(beginData[i], getSizeOfDimension(inputShape, i));
+        NN_RET_CHECK_LE(static_cast<uint32_t>(beginData[i]), getSizeOfDimension(inputShape, i));
         NN_RET_CHECK_GE(sliceSize, 0);
-        NN_RET_CHECK_LE(sliceBegin + sliceSize, getSizeOfDimension(inputShape, i));
+        NN_RET_CHECK_LE(static_cast<uint32_t>(sliceBegin + sliceSize),
+                        getSizeOfDimension(inputShape, i));
         outputShape.dimensions[i] = sliceSize;
     }
     return context->setOutputShape(kOutputTensor, outputShape);
